+++ CG184/src/Maths/Vector4D.h
@@ -7,38 +7,38 @@ namespace CG184
 	{
 	public:
 		Vector4D();
-		Vector4D(double _x, double _y, double _Z, double _w);
-		Vector4D(double val);
+		Vector4D(float _x, float _y, float _Z, float _w);
+		Vector4D(float val);
 		Vector4D(const Vector4D& _vec);
-		Vector4D(const Vector3D& _vec, double wval);
+		Vector4D(const Vector3D& _vec, float wval);
 
 		Vector4D add(const Vector4D& vec) const;
 		Vector4D sub(const Vector4D& vec) const;
 		Vector4D mult(const Vector4D& vec) const;
 		Vector4D divide(const Vector4D& vec) const;
 
-		Vector4D mult(double multiplier) const;
-		Vector4D divide(double diveder)const ;
+		Vector4D mult(float multiplier) const;
+		Vector4D divide(float diveder)const ;
 
-		const double at(size_t i) const;
-		double at(size_t i);
+		const float at(size_t i) const;
+		float at(size_t i);
 
 		bool isEqual(const Vector4D& vec) const;
 
-		double		operator[](size_t i) const;
+		float		operator[](size_t i) const;
 		Vector4D&	operator=(const Vector4D& vec);
 		Vector4D&   operator+(const Vector4D& vec);
 		Vector4D&   operator-(const Vector4D& vec);
 		Vector4D&   operator*(const Vector4D& vec);
 		Vector4D&   operator/(const Vector4D& vec);
-		Vector4D&   operator*(double val);
-		Vector4D&   operator/(double val);
+		Vector4D&   operator*(float val);
+		Vector4D&   operator/(float val);
 		Vector4D&   operator+=(const Vector4D& vec);
 		Vector4D&   operator-=(const Vector4D& vec);
 		Vector4D&   operator*=(const Vector4D& vec);
 		Vector4D&   operator/=(const Vector4D& vec);
-		Vector4D&   operator*=(double val);
-		Vector4D&   operator/=(double val);
+		Vector4D&   operator*=(float val);
+		Vector4D&   operator/=(float val);
 		bool		operator==(const Vector4D& vec);
 		bool		operator!=(const Vector4D& vec);
 
@@ -46,7 +46,7 @@ namespace CG184
 		~Vector4D();
 
 	public:
-		double x, y, z, w;
+		float x, y, z, w;
 	};
